+++ tango/integrations/beaker/executor.py
@@ -65,9 +65,9 @@ class ResourceAssignment(NamedTuple):
     Resources assigned to a step.
     """
 
-    cluster: str
+    cluster: Union[str, List[str]]
     """
-    The cluster to use to execute the step.
+    The cluster(s) to use to execute the step.
     """
 
     resources: TaskResources
@@ -127,7 +127,6 @@ class SimpleBeakerScheduler(BeakerScheduler):
         super().__init__()
         self.clusters = clusters
         self.priority = priority
-        self._latest_clusters_used: List[str] = []
         if not self.clusters:
             raise ConfigurationError("At least one cluster is required in 'clusters'")
 
@@ -139,44 +138,9 @@ class SimpleBeakerScheduler(BeakerScheduler):
             memory=step_resources.memory,
             shared_memory=step_resources.shared_memory,
         )
-        cluster_to_use = self._ensure_cluster(task_resources)
-        if cluster_to_use is None:
-            raise ResourceAssignmentError()
-
-        # Move cluster to the end of `self._latest_clusters_used`
-        try:
-            self._latest_clusters_used.remove(cluster_to_use)  # type: ignore
-        except ValueError:
-            pass
-        self._latest_clusters_used.append(cluster_to_use)  # type: ignore
-
         return ResourceAssignment(
-            cluster=cluster_to_use, resources=task_resources, priority=self.priority
+            cluster=self.clusters, resources=task_resources, priority=self.priority
         )
-
-    def _ensure_cluster(self, task_resources: TaskResources) -> Optional[str]:
-        cluster_to_use: Optional[str] = None
-        if not self.clusters:
-            raise ConfigurationError("At least one cluster is required in 'clusters'")
-        elif len(self.clusters) == 1:
-            cluster_to_use = self.clusters[0]
-        else:
-
-            def recency_ranking(cluster_name: str):
-                try:
-                    return self._latest_clusters_used.index(cluster_name)
-                except ValueError:
-                    return -1
-
-            available_clusters = sorted(
-                self.beaker.cluster.filter_available(task_resources, *self.clusters),
-                key=lambda x: (x.queued_jobs, recency_ranking(x.cluster.full_name)),
-            )
-
-            if available_clusters:
-                cluster_to_use = available_clusters[0].cluster.full_name
-
-        return cluster_to_use
 
 
 @Executor.register("beaker")
@@ -351,7 +315,7 @@ class BeakerExecutor(Executor):
         datasets: Optional[List[DataMount]] = None,
         env_vars: Optional[List[EnvVar]] = None,
         venv_name: Optional[str] = None,
-        parallelism: Optional[int] = -1,
+        parallelism: Optional[int] = None,
         install_cmd: Optional[str] = None,
         priority: Optional[Union[str, Priority]] = None,
         allow_dirty: bool = False,
@@ -606,9 +570,7 @@ class BeakerExecutor(Executor):
         update_steps_to_run()
 
         try:
-            with concurrent.futures.ThreadPoolExecutor(
-                max_workers=self.parallelism or None
-            ) as pool:
+            with concurrent.futures.ThreadPoolExecutor(max_workers=self.parallelism) as pool:
                 while steps_left_to_run:
                     # Submit steps left to run.
                     for step_name in steps_to_run:
@@ -733,17 +695,21 @@ class BeakerExecutor(Executor):
         assert experiment is not None
         assert experiment_url is not None
 
-        # Follow the experiment and stream the logs until it completes.
+        # Follow the experiment until it completes.
         try:
             while True:
                 try:
                     self._check_if_cancelled()
                     self.beaker.experiment.wait_for(
-                        experiment, strict=True, quiet=True, timeout=2.0
+                        experiment,
+                        strict=True,
+                        quiet=True,
+                        timeout=31,
+                        poll_interval=30,
                     )
-                    time.sleep(2.0)
                     break
                 except JobTimeoutError:
+                    time.sleep(30)
                     continue
         except (JobFailedError, TaskStoppedError):
             cli_logger.error(
@@ -896,7 +862,7 @@ class BeakerExecutor(Executor):
             )
 
         # Get cluster, resources, and priority to use.
-        cluster, task_resources, priority = self.scheduler.schedule(step)
+        clusters, task_resources, priority = self.scheduler.schedule(step)
         self._check_if_cancelled()
 
         # Ensure dataset with the entrypoint script exists and get it.
@@ -941,7 +907,6 @@ class BeakerExecutor(Executor):
         task_spec = (
             TaskSpec.new(
                 step.unique_id,
-                cluster,
                 beaker_image=self.beaker_image,
                 docker_image=self.docker_image,
                 result_path=Constants.RESULTS_DIR,
@@ -952,6 +917,7 @@ class BeakerExecutor(Executor):
                 env_vars=self.env_vars,
                 priority=priority,
             )
+            .with_constraint(cluster=[clusters] if isinstance(clusters, str) else clusters)
             .with_env_var(name="TANGO_VERSION", value=VERSION)
             .with_env_var(name="GITHUB_TOKEN", secret=Constants.GITHUB_TOKEN_SECRET_NAME)
             .with_env_var(name="BEAKER_TOKEN", secret=Constants.BEAKER_TOKEN_SECRET_NAME)